converso/pyproject.toml
# FILEPATH: /home/gianfranco/Desktop/uni/masters-thesis/converso/pyproject.toml

[tool.poetry]
name = "converso"
version = "0.1.0"
description = "Artificial Intelligence assistant"
authors = ["Gianfranco Demarco <gianfrademarco@gmail.com>"]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pygraphviz = "^1.12"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
pylint = "^3.0.3"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"
mkdocs-material = "^9.5.7"
mkdocstrings = {version = "^0.24.0", extras = ["python"]}
pymdown-extensions = "^10.7"
markdown-include = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10, <4.0"

aio_pika = "9.3.1"
fastapi = "0.108.0"
google-api-python-client = "2.111.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.0"
google-search-results = "2.4.2"
langserve = "0.0.36"
openai = "1.6.1"
pika = "1.3.2"
redis = "5.0.1"
sse_starlette = "1.8.2"
beautifulsoup4 = "4.12.2"
parsel = "1.8.1"
python-dateutil = "2.8.2"
readability-lxml = "0.8.1"
langgraph = "^0.0.15"
ipykernel = "^6.29.0"
langchain-openai = "^0.0.3"
langchainhub = "^0.1.14"
langchain = "^0.1.11"
langsmith = "^0.1.23"
grandalf = "^0.8"
